Refactor methods for localStorage
